OPSEC: hotelscontroller_cs--resolved-sql-injection-risks

+++ WebApplication452_simple/Controllers/HotelsController.cs
@@ -1,9 +1,9 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
+using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
 using WebApplication452_simple.Models;
@@ -15,24 +15,27 @@ namespace WebApplication452_simple.Controllers
         private ApplicationDbContext db = new ApplicationDbContext();
 
         // GET: Hotels
-        public ActionResult Index()
+        public async Task<ActionResult> Index()
         {
             var hotels = db.Hotels.Include(h => h.Country);
-            return View(hotels.ToList());
+            return View(await hotels.ToListAsync());
         }
 
         // GET: Hotels/Details/5
-        public ActionResult Details(int? id)
+        public async Task<ActionResult> Details(int? id)
         {
-            if (id == null)
+            if (!id.HasValue || id <= 0)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Hotel hotel = db.Hotels.Find(id);
+
+            var hotel = await db.Hotels.Include(h => h.Country)
+                                       .FirstOrDefaultAsync(h => h.Id == id);
             if (hotel == null)
             {
                 return HttpNotFound();
             }
+
             return View(hotel);
         }
 
@@ -44,16 +47,14 @@ namespace WebApplication452_simple.Controllers
         }
 
         // POST: Hotels/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Id,Name,CountryId,City,Stars,PricePerNight,IsAllInclusive")] Hotel hotel)
+        public async Task<ActionResult> Create([Bind(Include = "Id,Name,CountryId,City,Stars,PricePerNight,IsAllInclusive")] Hotel hotel)
         {
             if (ModelState.IsValid)
             {
                 db.Hotels.Add(hotel);
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
 
@@ -62,61 +63,74 @@ namespace WebApplication452_simple.Controllers
         }
 
         // GET: Hotels/Edit/5
-        public ActionResult Edit(int? id)
+        public async Task<ActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (!id.HasValue || id <= 0)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Hotel hotel = db.Hotels.Find(id);
+
+            var hotel = await db.Hotels.FirstOrDefaultAsync(h => h.Id == id);
             if (hotel == null)
             {
                 return HttpNotFound();
             }
+
             ViewBag.CountryId = new SelectList(db.Countries, "Id", "Name", hotel.CountryId);
             return View(hotel);
         }
 
         // POST: Hotels/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Id,Name,CountryId,City,Stars,PricePerNight,IsAllInclusive")] Hotel hotel)
+        public async Task<ActionResult> Edit([Bind(Include = "Id,Name,CountryId,City,Stars,PricePerNight,IsAllInclusive")] Hotel hotel)
         {
             if (ModelState.IsValid)
             {
                 db.Entry(hotel).State = EntityState.Modified;
-                db.SaveChanges();
+                await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
+
             ViewBag.CountryId = new SelectList(db.Countries, "Id", "Name", hotel.CountryId);
             return View(hotel);
         }
 
         // GET: Hotels/Delete/5
-        public ActionResult Delete(int? id)
+        public async Task<ActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (!id.HasValue || id <= 0)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Hotel hotel = db.Hotels.Find(id);
+
+            var hotel = await db.Hotels.FirstOrDefaultAsync(h => h.Id == id);
             if (hotel == null)
             {
                 return HttpNotFound();
             }
+
             return View(hotel);
         }
 
         // POST: Hotels/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            Hotel hotel = db.Hotels.Find(id);
+            if (id <= 0)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            var hotel = await db.Hotels.FirstOrDefaultAsync(h => h.Id == id);
+            if (hotel == null)
+            {
+                return HttpNotFound();
+            }
+
             db.Hotels.Remove(hotel);
-            db.SaveChanges();
+            await db.SaveChangesAsync();
             return RedirectToAction("Index");
         }
